Move mic level widget to separate component
This is highly desirable since it avoids a complete re-render when
the microphone level changes.
