+++ src/react-components/ui-root.js
@@ -4,7 +4,6 @@ import classNames from "classnames";
 import copy from "copy-to-clipboard";
 import { IntlProvider, FormattedMessage, addLocaleData } from "react-intl";
 import en from "react-intl/locale-data/en";
-import MovingAverage from "moving-average";
 import screenfull from "screenfull";
 
 import { VR_DEVICE_AVAILABILITY } from "../utils/vr-caps-detect";
@@ -57,7 +56,7 @@ import OAuthDialog from "./oauth-dialog.js";
 import LobbyChatBox from "./lobby-chat-box.js";
 import InWorldChatBox from "./in-world-chat-box.js";
 import AvatarEditor from "./avatar-editor";
-
+import MicLevelWidget from "./mic-level-widget.js";
 import PresenceLog from "./presence-log.js";
 import PresenceList from "./presence-list.js";
 import SettingsMenu from "./settings-menu.js";
@@ -202,9 +201,7 @@ class UIRoot extends Component {
     toneInterval: null,
     tonePlaying: false,
 
-    micLevel: 0,
     micDevices: [],
-    micUpdateInterval: null,
 
     profileNamePending: "Hello",
 
@@ -267,7 +264,6 @@ class UIRoot extends Component {
 
   componentDidMount() {
     window.addEventListener("concurrentload", this.onConcurrentLoad);
-    this.micLevelMovingAverage = MovingAverage(100);
     this.props.scene.addEventListener(
       "didConnectToNetworkedScene",
       () => {
@@ -585,39 +581,7 @@ class UIRoot extends Component {
     // we should definitely have an audioTrack at this point unless they denied mic access
     if (this.state.audioTrack) {
       mediaStream.addTrack(this.state.audioTrack);
-
-      const AudioContext = window.AudioContext || window.webkitAudioContext;
-      this.micLevelAudioContext = new AudioContext();
-      const micSource = this.micLevelAudioContext.createMediaStreamSource(mediaStream);
-      const analyser = this.micLevelAudioContext.createAnalyser();
-      analyser.fftSize = 32;
-      const levels = new Uint8Array(analyser.frequencyBinCount);
-
-      micSource.connect(analyser);
-
-      clearInterval(this.micUpdateInterval);
-      this.micUpdateInterval = setInterval(() => {
-        analyser.getByteTimeDomainData(levels);
-        let v = 0;
-        for (let x = 0; x < levels.length; x++) {
-          v = Math.max(levels[x] - 128, v);
-        }
-        const level = v / 128.0;
-        // Multiplier to increase visual indicator.
-        const multiplier = 6;
-        // We use a moving average to smooth out the visual animation or else it would twitch too fast for
-        // the css renderer to keep up.
-        this.micLevelMovingAverage.push(Date.now(), level * multiplier);
-        const average = this.micLevelMovingAverage.movingAverage();
-        this.setState(state => {
-          if (Math.abs(average - state.micLevel) > 0.0001) {
-            return { micLevel: average };
-          }
-        });
-      }, 50);
-
       const micDeviceId = this.micDeviceIdForMicLabel(this.micLabelForMediaStream(mediaStream));
-
       if (micDeviceId) {
         this.props.store.update({ settings: { lastUsedMicDeviceId: micDeviceId } });
       }
@@ -724,11 +688,6 @@ class UIRoot extends Component {
 
     this.stopTestTone();
     clearTimeout(this.testToneTimeout);
-
-    if (this.micLevelAudioContext) {
-      this.micLevelAudioContext.close();
-    }
-    clearInterval(this.micUpdateInterval);
 
     this.setState({ entered: true, showInviteDialog: false });
     clearHistoryState(this.props.history);
@@ -1188,9 +1147,6 @@ class UIRoot extends Component {
 
   renderAudioSetupPanel = () => {
     const maxLevelHeight = 111;
-    const micClip = {
-      clip: `rect(${maxLevelHeight - Math.floor(this.state.micLevel * maxLevelHeight)}px, 111px, 111px, 0px)`
-    };
     const speakerClip = { clip: `rect(${this.state.tonePlaying ? 0 : maxLevelHeight}px, 111px, 111px, 0px)` };
     const subtitleId = isMobilePhoneOrVR ? "audio.subtitle-mobile" : "audio.subtitle-desktop";
     return (
@@ -1210,40 +1166,7 @@ class UIRoot extends Component {
             {(isMobilePhoneOrVR || this.state.enterInVR) && <FormattedMessage id={subtitleId} />}
           </div>
           <div className="audio-setup-panel__levels">
-            <div className="audio-setup-panel__levels__icon">
-              <img
-                src="../assets/images/level_background.png"
-                srcSet="../assets/images/level_background@2x.png 2x"
-                className="audio-setup-panel__levels__icon-part"
-              />
-              {!this.state.muteOnEntry && (
-                <img
-                  src="../assets/images/level_fill.png"
-                  srcSet="../assets/images/level_fill@2x.png 2x"
-                  className="audio-setup-panel__levels__icon-part"
-                  style={micClip}
-                />
-              )}
-              {this.state.audioTrack && !this.state.muteOnEntry ? (
-                <img
-                  src="../assets/images/mic_level.png"
-                  srcSet="../assets/images/mic_level@2x.png 2x"
-                  className="audio-setup-panel__levels__icon-part"
-                />
-              ) : (
-                <img
-                  src="../assets/images/mic_denied.png"
-                  srcSet="../assets/images/mic_denied@2x.png 2x"
-                  className="audio-setup-panel__levels__icon-part"
-                />
-              )}
-              {this.state.audioTrack &&
-                !this.state.muteOnEntry && (
-                  <div className="audio-setup-panel__levels__test_label">
-                    <FormattedMessage id="audio.talk_to_test" />
-                  </div>
-                )}
-            </div>
+            <MicLevelWidget hasAudioTrack={!!this.state.audioTrack} muteOnEntry={this.state.muteOnEntry} mediaStream={this.state.mediaStream} />
             <WithHoverSound>
               <div className="audio-setup-panel__levels__icon_clickable" onClick={this.playTestTone}>
                 <img
